feat(shopping-cart): improve cart deletion and add payment toggle
Closes #923

Refactor deleteShoppingCart to use getOrCreateShoppingCart and
perform item and cart deletions concurrently for efficiency.

Add ButtonPricePickPaymentMethod component to toggle price display
between sunflower and euro units, enhancing payment method UI.

Introduce useShoppingCartDelete hook to handle cart deletion with
react-query mutation and cache invalidation.

Adjust ShoppingCartHud by removing unused imports and simplifying
code to improve maintainability.

Update NotificationList container max height for better UI layout.
