+++ packages/game/src/GameHud.tsx
@@ -15,20 +15,23 @@ import { RaisedBedFieldHud } from './hud/RaisedBedFieldHud';
 import { PaymentSuccessfulMessage } from './hud/PaymentSuccessfulMessage';
 import { GameSceneProps } from './GameScene';
 import { DebugHud } from './hud/DebugHud';
+import { useGameState } from './useGameState';
 
 export function GameHud({ flags }: { flags: GameSceneProps['flags'] }) {
+    const isCloseup = useGameState(state => state.view) === 'closeup';
+
     return (
         <>
             <div className='absolute top-2 left-2 flex flex-col items-start gap-2'>
                 <AccountHud />
-                <GameModeHud />
+                {!isCloseup && <GameModeHud />}
                 <ShoppingCartHud />
             </div>
             <div className='absolute top-2 right-2 flex items-end flex-col-reverse md:flex-row gap-1 md:gap-2'>
                 <WeatherHud />
                 <SunflowersHud />
             </div>
-            <DayNightCycleHud />
+            {!isCloseup && <DayNightCycleHud />}
             <div className='absolute bottom-0 flex flex-col left-0 right-0 md:flex-row md:justify-between md:items-end pointer-events-none'>
                 <div className='p-2 flex flex-row'>
                     <CameraHud />
