+++ packages/game/src/hud/raisedBed/RaisedBedCard.tsx
@@ -5,7 +5,7 @@ export function RaisedBedCard({ className, ...rest }: HTMLAttributes<HTMLDivElem
     return (
         <div
             className={cx(
-                'bg-gradient-to-br from-lime-100/90 dark:from-lime-200/80 to-lime-100/80 dark:to-lime-200/70 dark:text-green-950 rounded-3xl',
+                'bg-gradient-to-br from-lime-100/90 dark:from-lime-200/80 to-lime-100/80 dark:to-lime-200/70 dark:text-green-950 rounded-sm md:rounded-3xl',
                 className
             )}
             {...rest} />
